@@ -15,14 +15,13 @@ classifiers =
   Intended Audience :: Science/Research
   License :: OSI Approved :: LGPL
   Programming Language :: Python
-  Programming Language :: Python :: 3.7
-  Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
+  Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3 :: Only
   Framework :: Jupyter
 
 [options]
-python_requires = >=3.7
+python_requires = >=3.9
 
 zip_safe = False
 packages = find:
@@ -31,8 +30,11 @@ install_requires =
   ipykernel>=4.5.1
   ipython>=6.1.0
   jupyterlab>3.2.0
-  ipywidgets>=7.6.5,<8.0
+  jupyterlab_server
+  ipywidgets>=7.7.2,<8.0
   PyYAML>=5.4
-  z3-solver>=4.8.15
+  z3-solver>=4.8.15,<4.12
   networkx>=2.8
   matplotlib>3.5
+  netcdf4>=1.6
+  xarray>=0.20
