@@ -1,9 +1,9 @@
 from IPython.display import display, HTML
 
 def alert_warning(msg):
-    js = """<script>alert("WARNING: {} ");</script>""".format(msg)
+    js = f"""<script>alert("WARNING: {msg} ");</script>"""
     display(HTML(js))
 
 def alert_error(msg):
-    js = "<script>alert('ERROR: {}');</script>".format(msg)
+    js = f"<script>alert('ERROR: {msg}');</script>"
     display(HTML(js))
