@@ -1,39 +1,46 @@
 import logging
-from traitlets import Unicode, validate
-from z3 import SeqRef, main_ctx, Z3_mk_const, to_symbol, StringSort
+from traitlets import Bool, validate
+from z3 import BoolRef, main_ctx, Z3_mk_const, to_symbol, BoolSort
 
 from visualCaseGen.OutHandler import handler as owh
 from visualCaseGen.config_var import ConfigVar
 from visualCaseGen.dialog import alert_error
 from visualCaseGen.logic import logic
+from specs.options_specs import OptionsSpec
 
 logger = logging.getLogger("\t" + __name__.split(".")[-1])
 
 
-class ConfigVarStr(ConfigVar, SeqRef):
-    """A derived ConfigVar class with value of type String. Each instance can have a single
-    string or None as its value.
-
-    self._widget.value  : value preceded by a validity char. (of type String -- Trait)
-    self.value:         : value NOT preceded by a validity char. (of type String -- Trait)
+class ConfigVarBool(ConfigVar, BoolRef):
+    """A derived ConfigVar class with value of type Bool.
     """
 
     # trait
-    value = Unicode(allow_none=True)
+    value = Bool(allow_none=True)
 
     def __init__(self, name, *args, **kwargs):
 
-        # Initialize the ConfigVarsuper class
+        # Initialize the ConfigVar super class
         ConfigVar.__init__(self, name, *args, **kwargs)
 
-        # Initialize the SeqRef super class, i.e., a Z3 string
-        # Below initialization mimics String() definition in z3.py
+        # Initialize the BoolRef super class, i.e., a Z3 bool
+        # Below initialization mimics Bool() definition in z3.py
         ctx = main_ctx()
-        SeqRef.__init__(
-                self, 
-                Z3_mk_const(ctx.ref(), to_symbol(name, ctx), StringSort(ctx).ast),
-                ctx
+        BoolRef.__init__(
+            self,
+            Z3_mk_const(ctx.ref(), to_symbol(name, ctx), BoolSort(ctx).ast),
+            ctx
+        )
+
+        # Create an OptionsSpec instance for each ConfigVarBool object,
+        # so as to make sure that they have a finite options list: [True, False]
+        OptionsSpec(
+            var = self,
+            options_and_tooltips = (
+                [True, False],
+                [True, False]
             )
+        )
 
     @validate("value")
     def _validate_value(self, proposal):
@@ -45,14 +52,13 @@ class ConfigVarStr(ConfigVar, SeqRef):
 
         # confirm the value validity
         if self.has_finite_options_list():
-            try:
-                if self._options_validities[new_val] is False:
-                    err_msg = logic.retrieve_error_msg(self, new_val)
-                    raise AssertionError(err_msg)
-            except KeyError:
-                raise AssertionError(f"{new_val} not an option for {self.name}")
+            if self._options_validities[new_val] is False:
+                err_msg = logic.retrieve_error_msg(self, new_val)
+                raise AssertionError(err_msg)
 
-        else: # infinite domain
+        else:
+            # actually, this should never be reached since boolean vars always have finite
+            # list of options.
             logic.check_assignment(self, new_val)
 
         # finally, set self.value by returning new_vals
@@ -63,11 +69,10 @@ class ConfigVarStr(ConfigVar, SeqRef):
         """This methods gets called by _post_value_change and other methods to update the
         displayed widget value whenever the internal value changes. In other words, this
         method propagates backend value change to frontend."""
-
-        if self.value is None or not self.has_finite_options_list():
+        if self.value is None:
             self._widget.value = self._widget_none_val
         else:
-            self._widget.value = self._valid_opt_char + " " + self.value
+            self._widget.value = self._valid_opt_char + " " + str(self.value)
 
     @owh.out.capture()
     def _process_frontend_value_change(self, change):
@@ -85,6 +90,14 @@ class ConfigVarStr(ConfigVar, SeqRef):
             new_widget_val = change["owner"].value
             new_val_validity_char = new_widget_val[0]
             new_val = new_widget_val[1:].strip()
+
+            if new_val == "True":
+                new_val = True
+            elif new_val == "False":
+                new_val = False
+            else:
+                raise RuntimeError(f"Encountered an invalid value type for ConfigVarBool instance {self.name}")
+
 
             if new_widget_val == self._widget_none_val and self._always_set is True:
                 # Attempted to set the value to None while always_set property is True.
