@@ -1,34 +1,34 @@
 import logging
-from traitlets import Int, validate
-from z3 import ArithRef, main_ctx, Z3_mk_const, to_symbol, IntSort
+from traitlets import Float, validate
+from z3 import ArithRef, main_ctx, Z3_mk_const, to_symbol, RealSort
 
 from visualCaseGen.OutHandler import handler as owh
 from visualCaseGen.config_var import ConfigVar
 from visualCaseGen.dialog import alert_error
 from visualCaseGen.logic import logic
-from visualCaseGen.dev_utils import is_integer
+from visualCaseGen.dev_utils import is_number
 
 logger = logging.getLogger("\t" + __name__.split(".")[-1])
 
 
-class ConfigVarInt(ConfigVar, ArithRef):
-    """A derived ConfigVar class with value of type Int.
+class ConfigVarReal(ConfigVar, ArithRef):
+    """A derived ConfigVar class with value of type Real.
     """
 
     # trait
-    value = Int(allow_none=True)
+    value = Float(allow_none=True)
 
     def __init__(self, name, *args, **kwargs):
 
         # Initialize the ConfigVar super class
         ConfigVar.__init__(self, name, *args, **kwargs, widget_none_val='')
 
-        # Initialize the ArithRef super class, i.e., a Z3 int
-        # Below initialization mimics Int() definition in z3.py
+        # Initialize the ArithRef super class, i.e., a Z3 Real
+        # Below initialization mimics Real() definition in z3.py
         ctx = main_ctx()
         ArithRef.__init__(
                 self,
-                Z3_mk_const(ctx.ref(), to_symbol(name, ctx), IntSort(ctx).ast),
+                Z3_mk_const(ctx.ref(), to_symbol(name, ctx), RealSort(ctx).ast),
                 ctx
             )
 
@@ -76,9 +76,9 @@ class ConfigVarInt(ConfigVar, ArithRef):
                 self.value = None
                 return
 
-            # First, check if the string is an integer
-            if not is_integer(new_val):
-                alert_error(f"The value {new_val} entered for {self.name} is not an integer!")
+            # First check if the string is a number
+            if not is_number(new_val):
+                alert_error(f"The value {new_val} entered for {self.name} is not a number!")
                 # set back to old value:
                 self._widget.value = self.widget_none_val if self.value is None else str(self.value)
                 return
@@ -89,12 +89,12 @@ class ConfigVarInt(ConfigVar, ArithRef):
             if outcome is False:
                 logger.critical("ERROR: %s", err_msg)
                 alert_error(err_msg)
-                # set back to old value:
+                # set to old value:
                 self._widget.value = self.widget_none_val if self.value is None else str(self.value)
                 return
 
             else:
-                self.value = int(new_val)
+                self.value = float(new_val)
 
 
     @ConfigVar.widget.setter
