@@ -1,6 +1,8 @@
 import logging
 from visualCaseGen.config_var import ConfigVar
+from visualCaseGen.config_var_bool import ConfigVarBool
 from visualCaseGen.config_var_int import ConfigVarInt
+from visualCaseGen.config_var_real import ConfigVarReal
 from visualCaseGen.config_var_str import ConfigVarStr
 from visualCaseGen.config_var_str_ms import ConfigVarStrMS
 from visualCaseGen.config_var_compset import ConfigVarCompset
@@ -11,7 +13,6 @@ def init_configvars(ci, predefined_mode=False):
     """ Define the ConfigVars and, by doing so, register them with the logic engine. """
     logger.debug("Initializing ConfigVars...")
 
-    ConfigVarInt('ILI')
     ConfigVarStr('INITTIME')
     for comp_class in ci.comp_classes:
         ConfigVarStr('COMP_'+str(comp_class))
@@ -30,4 +31,29 @@ def init_configvars(ci, predefined_mode=False):
     ConfigVarStr('MASK_GRID')
     cv_grid = ConfigVarStrMS('GRID')
     cv_grid.view_mode = 'suggested' # or 'all'
+
+    ConfigVarStr('GRID_MODE') # 'Predefined' | 'Custom'
+    
+    ### Custom ocean grid variables
+    ConfigVarStr('OCN_GRID_EXTENT') # 'Regional' | 'Global'
+    ConfigVarStr('OCN_GRID_CONFIG')
+    ConfigVarStr('OCN_AXIS_UNITS')
+    ConfigVarBool('OCN_CYCLIC_X')
+    ConfigVarBool('OCN_CYCLIC_Y')
+    ConfigVarBool('OCN_TRIPOLAR_N')
+    ConfigVarInt('OCN_NX')
+    ConfigVarInt('OCN_NY')
+    ConfigVarReal('OCN_LENX')
+    ConfigVarReal('OCN_LENY')
+
+    ### Custom land grid variables
+    #ConfigVarReal('LND_LAT_1')
+    #ConfigVarReal('LND_LAT_2')
+    #ConfigVarReal('LND_LON_1')
+    #ConfigVarReal('LND_LON_2')
+    ConfigVarInt('LND_DOM_PFT')
+    ConfigVarInt('LND_SOIL_COLOR')
+    ConfigVarReal('LND_MAX_SAT_AREA')
+    ConfigVarReal('LND_STD_ELEV')
+
     ConfigVar.lock()
