@@ -13,9 +13,6 @@ from z3 import BoolRef, Int
 from z3 import z3util
 import networkx as nx
 
-import cProfile, pstats
-profiler = cProfile.Profile()
-
 logger = logging.getLogger('\t\t'+__name__.split('.')[-1])
 
 class Logic():
@@ -261,6 +258,8 @@ class Logic():
                 if var.has_options_spec() and varname=="GRID"] 
         else:
             relational_assertions = relational_assertions_setter(cvars)
+            if len(set(relational_assertions.values())) != len(relational_assertions.values()):
+                raise RuntimeError("Error messages associated with each relational assertion must be identical.")
             options_specs = [var._options_spec for varname, var in cvars.items() \
                 if var.has_options_spec()]
         
@@ -284,8 +283,8 @@ class Logic():
             layer.add_asrt_option(var, new_opts)
 
     @classmethod
-    def check_assignment(cls, var, new_value):
-        var.major_layer.check_assignment(var, new_value)
+    def check_assignment(cls, var, new_value, return_outcome=False):
+        return var.major_layer.check_assignment(var, new_value, return_outcome)
 
     @classmethod
     def get_options_validities(cls, var):
@@ -348,7 +347,7 @@ class Layer():
 
         # Set of variables whose options validities are to be updated only.
         # These are variables whose neighbors went through either a value change or an options validities change
-        self.vars_refresh_validities = list()
+        self.vars_refresh_validities = set()
         # set of variables whose options are to be updated.
         # These are variables who are options children of variables whose values are changed.
         self.vars_refresh_options = set()
@@ -412,15 +411,14 @@ class Layer():
         as variables whose options/validities are to be potentially updated."""
 
         major_layer = var.major_layer
-        major_layer.vars_refresh_validities.extend(
-            [var_other for var_other in var.peer_vars_relational if var_other not in major_layer.vars_refresh_validities]            
+        major_layer.vars_refresh_validities.update(
+            [var_other for var_other in var.peer_vars_relational]
         )
 
         # update set of child vars whose validities are to be refreshed:
         for child_var_rlt in var.child_vars_relational:
             child_layer = child_var_rlt.major_layer
-            if child_var_rlt not in child_layer.vars_refresh_validities:
-                child_layer.vars_refresh_validities.append(child_var_rlt)
+            child_layer.vars_refresh_validities.add(child_var_rlt)
 
         # update set of variables whose options are to be updated due to the value change of this var:
         if designate_opt_children:
@@ -435,7 +433,7 @@ class Layer():
         self._solver.pop()
         return new_validities
 
-    def check_assignment(self, var, new_value):
+    def check_assignment(self, var, new_value, return_outcome=False):
 
         status = True
         err_msg = ''
@@ -457,7 +455,9 @@ class Layer():
                 if status is False:
                     err_msg = self.retrieve_error_msg(var, new_value)
 
-        if status is False:
+        if return_outcome:
+            return status, err_msg
+        elif status is False:
             raise AssertionError(err_msg)
 
     def retrieve_error_msg(self, var, value):
@@ -496,10 +496,8 @@ class Layer():
         if not some_relational_impact():
             return
 
-        ivar = 0
-        while len(self.vars_refresh_validities) > ivar:
-            var = self.vars_refresh_validities[ivar]
-            ivar += 1
+        while len(self.vars_refresh_validities) > 0:
+            var = self.vars_refresh_validities.pop()
             if var.has_options():
 
                 # Determine new validities
@@ -521,8 +519,6 @@ class Layer():
                             format(var.name, var.value, invoker_var.name, invoker_var.value))
                     var.update_options_validities(new_validities=new_validities)
         
-        self.vars_refresh_validities.clear()
-
     
     def _sweep_subsq_layer(self, invoker_var):
 
@@ -600,10 +596,8 @@ class Layer():
         self.vars_refresh_options.clear()
 
         # refresh validities
-        ivar = 0
-        while len(self.vars_refresh_validities) > ivar:
-            var = self.vars_refresh_validities[ivar]
-            ivar += 1 
+        while len(self.vars_refresh_validities) > 0:
+            var = self.vars_refresh_validities.pop()
             if var.has_options():
                 # Determine new validities
                 s.push()
@@ -623,7 +617,6 @@ class Layer():
                         raise RunError("The {}={} assignment is no longer valid after {}={} assignment!".\
                             format(var.name, var.value, invoker_var.name, invoker_var.value))
                     var.update_options_validities(new_validities=new_validities)
-        self.vars_refresh_validities.clear()
 
         # set values of reset_variables, if need be
         for var in reset_variables:
@@ -631,10 +624,8 @@ class Layer():
                 var.value = var.get_first_valid_option()
 
         # refresh validities again if need be. 
-        ivar = 0
-        while len(self.vars_refresh_validities) > ivar:
-            var = self.vars_refresh_validities[ivar]
-            ivar += 1 
+        while len(self.vars_refresh_validities) > 0:
+            var = self.vars_refresh_validities.pop()
             if var.has_options():
                 # Determine new validities
                 s.push()
@@ -654,7 +645,6 @@ class Layer():
                         raise RunError("The {}={} assignment is no longer valid after {}={} assignment!".\
                             format(var.name, var.value, invoker_var.name, invoker_var.value))
                     var.update_options_validities(new_validities=new_validities)
-        self.vars_refresh_validities.clear()
 
         # finally, inform the user about the indirect value changes occured in this layer
         # due to the invoker value change:
